+++ app/src/main/java/de/triology/cas/services/RegistryEtcd.java
@@ -40,6 +40,26 @@ class RegistryEtcd implements Registry {
     private final EtcdClient etcd;
 
     private static final String CAS_SERVICE_ACCOUNT_DIR = "/config/cas/service_accounts";
+
+    private enum CasServiceAccountTypes {
+        OAUTH("oauth"),
+        OIDC("oidc"),
+        UNDEFINED("");
+
+        private final String id;
+
+        CasServiceAccountTypes(String id) {
+            this.id = id;
+        }
+
+        public static CasServiceAccountTypes getByString(String id) {
+            for (CasServiceAccountTypes e : values()) {
+                if (e.id.equals(id)) return e;
+            }
+            return UNDEFINED;
+        }
+    }
+
     public static final String SERVICE_ACCOUNT_TYPE_OAUTH = "oauth";
     public static final String SERVICE_ACCOUNT_TYPE_OIDC = "oidc";
 
@@ -58,18 +78,16 @@ class RegistryEtcd implements Registry {
     public List<CesServiceData> getInstalledCasServiceAccountsOfType(String type, ICesServiceFactory factory) {
         log.debug("Get [{}] service accounts from registry", type);
         try {
-            List<EtcdKeysResponse.EtcdNode> nodes = etcd.getDir(CAS_SERVICE_ACCOUNT_DIR + "/" + type).send().get().getNode().getNodes();
+            List<EtcdKeysResponse.EtcdNode> nodes = etcd.getDir(String.format("%s/%s", CAS_SERVICE_ACCOUNT_DIR, type)).send().get().getNode().getNodes();
             return extractServiceAccountClientsByType(nodes, type, factory);
         } catch (EtcdException e) {
             if (e.isErrorCode(EtcdErrorCode.KeyNotFound)) {
                 return new ArrayList<>();
             } else {
-                log.warn("Failed to getInstalledCasServiceAccountsOfType: ", e);
-                throw new RegistryException(e);
+                throw new RegistryException("Failed to getInstalledCasServiceAccountsOfType: ", e);
             }
         } catch (IOException | EtcdAuthenticationException | TimeoutException e) {
-            log.error("Failed to getInstalledCasServiceAccountsOfType: ", e);
-            throw new RegistryException(e);
+            throw new RegistryException("Failed to getInstalledCasServiceAccountsOfType: ", e);
         }
     }
 
@@ -85,14 +103,14 @@ class RegistryEtcd implements Registry {
         List<CesServiceData> serviceDataList = new ArrayList<>();
         for (EtcdKeysResponse.EtcdNode oAuthClient : nodesFromEtcd) {
             try {
-                String clientPathPrefix = CAS_SERVICE_ACCOUNT_DIR + "/" + type + "/";
-                String clientID = oAuthClient.getKey().substring(clientPathPrefix.length());
+                var clientPathPrefix = String.format("%s/%s/", CAS_SERVICE_ACCOUNT_DIR, type);
+                var clientID = oAuthClient.getKey().substring(clientPathPrefix.length());
                 HashMap<String, String> attributes = new HashMap<>();
 
-                switch (type) {
-                    case SERVICE_ACCOUNT_TYPE_OIDC:
-                    case SERVICE_ACCOUNT_TYPE_OAUTH:
-                        String clientSecret = getEtcdValueForKeyIfPresent(clientPathPrefix + clientID);
+                switch (CasServiceAccountTypes.getByString(type)) {
+                    case OIDC:
+                    case OAUTH:
+                        var clientSecret = getEtcdValueForKeyIfPresent(clientPathPrefix + clientID);
                         attributes.put(CesOAuthServiceFactory.ATTRIBUTE_KEY_OAUTH_CLIENT_ID, clientID);
                         attributes.put(CesOAuthServiceFactory.ATTRIBUTE_KEY_OAUTH_CLIENT_SECRET_HASH, clientSecret);
                         break;
@@ -102,7 +120,7 @@ class RegistryEtcd implements Registry {
 
                 serviceDataList.add(new CesServiceData(clientID, factory, attributes));
             } catch (RegistryException ex) {
-                log.error("registry exception occurred", ex);
+                throw new RuntimeException("registry exception occurred", ex);
             }
         }
         return serviceDataList;
@@ -114,15 +132,15 @@ class RegistryEtcd implements Registry {
         for (EtcdKeysResponse.EtcdNode dogu : nodesFromEtcd) {
             JSONObject json;
             try {
-                String doguName = dogu.getKey().substring(DOGU_DIR.length() + 1);
+                var doguName = dogu.getKey().substring(DOGU_DIR.length() + 1);
                 json = getCurrentDoguNode(doguName);
                 if (hasCasDependency(json)) {
                     doguServices.add(new CesServiceData(doguName, factory));
                 }
             } catch (ParseException ex) {
-                log.error("failed to parse EtcdNode to json: ", ex);
+                throw new RuntimeException("failed to parse EtcdNode to json: ", ex);
             } catch (RegistryException ex) {
-                log.error("registry exception occurred: ", ex);
+                throw new RuntimeException("registry exception occurred: ", ex);
             }
         }
         return doguServices;
@@ -135,8 +153,7 @@ class RegistryEtcd implements Registry {
             List<EtcdKeysResponse.EtcdNode> nodes = etcd.getDir(DOGU_DIR).send().get().getNode().getNodes();
             return extractDogusFromDoguRootDir(nodes, factory);
         } catch (IOException | EtcdException | EtcdAuthenticationException | TimeoutException e) {
-            log.error("Failed to getInstalledDogusWhichAreUsingCAS: ", e);
-            throw new RegistryException(e);
+            throw new RegistryException("Failed to getInstalledDogusWhichAreUsingCAS: ", e);
         }
     }
 
@@ -150,15 +167,9 @@ class RegistryEtcd implements Registry {
         try {
             return etcd.get(key).send().get().getNode().getValue();
         } catch (EtcdException e) {
-            if (e.isErrorCode(EtcdErrorCode.KeyNotFound)) {
-                log.error("Failed to getEtcdValueForKey: {}", key);
-            } else {
-                log.error("Failed to getEtcdValueForKey: ", e);
-            }
-            throw new RegistryException(e);
+            throw new RegistryException(String.format("Failed to getEtcdValueForKey: %s", key), e);
         } catch (IOException | EtcdAuthenticationException | TimeoutException e) {
-            log.error("Failed to getEtcdValueForKey: ", e);
-            throw new RegistryException(e);
+            throw new RegistryException("Failed to getEtcdValueForKey: ", e);
         }
     }
 
@@ -178,12 +189,10 @@ class RegistryEtcd implements Registry {
                 //Valid case if key is not found return an empty string
                 return "";
             } else {
-                log.warn("Failed to getEtcdValueForKey: ", e);
-                throw new RegistryException(e);
+                throw new RegistryException("Failed to getEtcdValueForKey: ", e);
             }
         } catch (IOException | EtcdAuthenticationException | TimeoutException e) {
-            log.warn("Failed to getEtcdValueForKey: ", e);
-            throw new RegistryException(e);
+            throw new RegistryException("Failed to getEtcdValueForKey: ", e);
         }
     }
 
@@ -241,8 +250,7 @@ class RegistryEtcd implements Registry {
                     doguChangeListener.onChange();
                 }
             } catch (IOException | EtcdException | TimeoutException | EtcdAuthenticationException e) {
-                log.error("Failed to addDoguChangeListener: ", e);
-                throw new RegistryException(e);
+                throw new RegistryException("Failed to addDoguChangeListener for dogus: ", e);
             }
         });
         t.start();
@@ -256,8 +264,7 @@ class RegistryEtcd implements Registry {
                     doguChangeListener.onChange();
                 }
             } catch (IOException | EtcdException | TimeoutException | EtcdAuthenticationException e) {
-                log.error("Failed to addDoguChangeListener: ", e);
-                throw new RegistryException(e);
+                throw new RegistryException("Failed to addDoguChangeListener for service accounts: ", e);
             }
         });
         t2.start();
@@ -270,10 +277,10 @@ class RegistryEtcd implements Registry {
     protected JSONObject getCurrentDoguNode(String doguName) throws ParseException {
         JSONObject json = null;
         // get used dogu version
-        String doguVersion = getEtcdValueForKeyIfPresent(DOGU_DIR + "/" + doguName + "/current");
+        String doguVersion = getEtcdValueForKeyIfPresent(String.format("%s/%s/current", DOGU_DIR, doguName));
         // empty if dogu isnt used
         if (!doguVersion.isEmpty()) {
-            String doguDescription = getEtcdValueForKey(DOGU_DIR + "/" + doguName + "/" + doguVersion);
+            String doguDescription = getEtcdValueForKey(String.format("%s/%s/%s", DOGU_DIR, doguName, doguVersion));
             json = (JSONObject) PARSER.parse(doguDescription);
         }
         return json;